+++ GUI/.eric6project/receivemessageGUI.e6t
@@ -1,16 +1,16 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE Tasks SYSTEM "Tasks-6.0.dtd">
 <!-- eric6 tasks file for project receivemessageGUI -->
-<!-- Saved: 2021-05-31, 15:21:23 -->
+<!-- Saved: 2021-06-04, 10:50:03 -->
 <Tasks version="6.0">
   <ProjectScanFilter></ProjectScanFilter>
-  <Task priority="1" completed="False" type="1" uid="{7bd3fc50-1354-4a0b-acab-20b6ea3a8a1a}">
+  <Task priority="1" completed="False" type="1" uid="{11aeb1a0-3c8b-4564-ac3d-f15ec46d97d4}">
     <Summary>TODO: not implemented yet</Summary>
     <Description></Description>
-    <Created>2021-05-31, 14:28:53</Created>
+    <Created>2021-06-04, 10:50:00</Created>
     <Resource>
-      <Filename>ui/mainwindow.py</Filename>
-      <Linenumber>42</Linenumber>
+      <Filename>receivemessageGUI.py</Filename>
+      <Linenumber>185</Linenumber>
     </Resource>
   </Task>
 </Tasks>
