--- GUI/receivemessageengine.py
@@ -1,10 +1,0 @@
-from ui.mainwindow import MainWindow
-#from time import sleep
-
-#class Engine():
-    #def test(self):
-#   self.maxTemp.setText('11')
-def label_update():
-    print('label_update')
-    #sleep(2)
-    MainWindow().update_label()
